ml/analysis/requirements.txt
annotated-types==0.7.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
filelock==3.16.1
Flask==3.0.2
fsspec==2024.12.0
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.157.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
grpcio==1.69.0
grpcio-status==1.69.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httplib2==0.22.0
huggingface-hub==0.27.0
Hypercorn==0.17.3
hyperframe==6.0.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
packaging==24.2
pillow==10.2.0
priority==2.0.0
proto-plus==1.25.0
protobuf==5.29.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.4
pydantic_core==2.27.2
pyparsing==3.2.1
python-dotenv==1.0.1
python-json-logger==2.0.7
PyYAML==6.0.2
regex==2024.11.6
requests==2.31.0
rsa==4.9
safetensors==0.4.5
scikit-learn==1.6.0
scipy==1.12.0
sentence-transformers==2.6.1
sympy==1.13.3
threadpoolctl==3.5.0
tokenizers==0.15.2
torch==2.2.2
tqdm==4.67.1
transformers==4.38.2
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.29.0
Werkzeug==3.1.3
wsproto==1.2.0
